# perf(logging): reduce peak memory usage in CDF export