--- a/libs/logging/src/index.ts
+++ b/libs/logging/src/index.ts
@@ -1,8 +1,9 @@
 /* istanbul ignore file */
-export * from './logger';
-export * from './types';
+export * from './export';
+export * from './helpers';
 export * from './log_event_ids';
 export * from './log_event_types';
 export * from './log_source';
+export * from './logger';
 export * from './test_utils';
-export * from './helpers';
+export * from './types';